RavenDB-24887 - RavenDB-24887: Move ExecuteMultiAgentAndQueryRequestsAsync and related methods and class into ConversationHandler.ExecuteMultiAgentAndQueryRequestsAsync partial class
